converter: merge module inside of workflows/utils
As all the remaining uses of this module are in the ``workflows``
module, we might as well make it part of its utils.
