+++ inspirehep/modules/workflows/tasks/arxiv.py
@@ -39,7 +39,7 @@ from dojson.contrib.marc21.utils import create_record
 
 from inspire_dojson.hep import hep
 from inspire_schemas.utils import classify_field
-from inspirehep.modules.converter import convert
+from inspirehep.modules.workflows.utils import convert
 from inspirehep.utils.record import get_arxiv_categories, get_arxiv_id
 from inspirehep.utils.url import is_pdf_link
