+++ inspirehep/modules/workflows/utils/stylesheets/authorlist2marcxml.xsl
